make contact method a react component
